check prev_file_str before write op

+++ cvebased/repo.py
@@ -47,7 +47,8 @@ def add_cve_front_matter(path_to_repo: str, data: dict) -> None:
     if exists:
         # if file exists, parse existing data and overwrite file
         with open(path_to_cve, 'r') as file:
-            ex_data, advisory = parse_md(file.read())
+            file_str = file.read()
+            ex_data, advisory = parse_md(file_str)
             file.close()
         del data['id']
         for k, v in data.items():
@@ -55,22 +56,39 @@ def add_cve_front_matter(path_to_repo: str, data: dict) -> None:
                 ex_data.setdefault(k, []).append(el)
             if len(ex_data[k]) > 1:
                 ex_data[k] = dedupe_sort(ex_data[k])
-        write_md(path_to_cve, ex_data, advisory)
+        write_md(path_to_cve, ex_data, advisory, file_str)
     else:
         # if not exist, generate new file
         compile_cve(path_to_repo, data)
 
 
-def write_md(filepath, front_matter, markdown=''):
+def write_md(
+        filepath: AnyStr,
+        front_matter: Dict,
+        markdown: Optional[AnyStr] = '',
+        prev_file_str: Optional[AnyStr] = ''
+) -> bool:
+    """Writes front matter & markdown to a given filepath.
+
+    Skips file write operation if previous file content (passed in as string) is unchanged.
+
+    Returns:
+        True or False whether file write occurred."""
+
+    write_str = "---\n"
+    write_str += object_to_yaml_str(front_matter)
+    write_str += "---\n"
+    if markdown != '':
+        write_str += f"{markdown}\n"
+    # exit without IO operation if no changes in file content
+    if prev_file_str != '' and write_str == prev_file_str:
+        return False
     with open(filepath, 'w+') as file:
         file.seek(0)
-        file.write("---\n")
-        yaml.dump(front_matter, file)
-        file.write("---\n")
-        if markdown != '':
-            file.write("{}\n".format(markdown))
+        file.write(write_str)
         file.truncate()
         file.close()
+        return True
 
 
 def parse_md(content: str) -> (dict, str):
